refactor(Kafka Trigger Node): Split into versioned v1 module
Move the Kafka Trigger's single flat class into the folder-per-version
layout n8n uses for multi-version nodes, following the same split already
done for the sibling Kafka action node. The entry file becomes a thin
VersionedNodeType wrapper mapping typeVersions 1, 1.1, 1.2, and 1.3 to the
new KafkaTriggerV1 class; all trigger logic moves unchanged. defaultVersion
is now set explicitly to 1.3 instead of relying on the framework picking
the highest version in the array, so a future version 2 does not silently
become the default for new workflows.
