+++ packages/nodes-base/nodes/Kafka/test/KafkaTrigger.node.test.ts
@@ -12,11 +12,18 @@ import {
 	type RecordBatchEntry,
 	type RetryOptions,
 } from 'kafkajs';
-import { NodeOperationError, TriggerCloseError, UnexpectedError, type IRun } from 'n8n-workflow';
+import {
+	NodeOperationError,
+	TriggerCloseError,
+	UnexpectedError,
+	type INodeTypeBaseDescription,
+	type IRun,
+} from 'n8n-workflow';
 
 import { testTriggerNode } from '@test/nodes/TriggerHelpers';
 
 import { KafkaTrigger } from '../KafkaTrigger.node';
+import { KafkaTriggerV1 } from '../v1/KafkaTriggerV1.node';
 import type { Mock, Mocked } from 'vitest';
 
 vi.mock('kafkajs');
@@ -24,6 +31,15 @@ vi.mock('@kafkajs/confluent-schema-registry');
 vi.mock('@n8n/utils/sleep', () => ({
 	sleep: vi.fn().mockResolvedValue(undefined),
 }));
+
+const baseDescription: INodeTypeBaseDescription = {
+	displayName: 'Kafka Trigger',
+	name: 'kafkaTrigger',
+	icon: { light: 'file:kafka.svg', dark: 'file:kafka.dark.svg' },
+	group: ['trigger'],
+	defaultVersion: 1.3,
+	description: 'Consume messages from a Kafka topic',
+};
 
 describe('KafkaTrigger Node', () => {
 	const expectedRetryConfig = {
@@ -175,7 +191,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should connect to Kafka and subscribe to topic', async () => {
-		const { close, emit } = await testTriggerNode(KafkaTrigger, {
+		const { close, emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1,
@@ -254,7 +270,7 @@ describe('KafkaTrigger Node', () => {
 		};
 
 		const startTrigger = async () =>
-			await testTriggerNode(KafkaTrigger, {
+			await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -312,7 +328,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should handle authentication when credentials are provided', async () => {
-		await testTriggerNode(KafkaTrigger, {
+		await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -347,7 +363,7 @@ describe('KafkaTrigger Node', () => {
 
 	it('should throw an error if authentication is enabled but credentials are missing', async () => {
 		await expect(
-			testTriggerNode(KafkaTrigger, {
+			testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					parameters: {
@@ -366,7 +382,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use schema registry when enabled', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -407,7 +423,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use the schema registry credential when selected', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				credentials: {
@@ -460,7 +476,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use the schema registry credential without auth when authentication is none', async () => {
-		await testTriggerNode(KafkaTrigger, {
+		await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				credentials: {
@@ -504,7 +520,7 @@ describe('KafkaTrigger Node', () => {
 			),
 		);
 
-		const { emit, logger } = await testTriggerNode(KafkaTrigger, {
+		const { emit, logger } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -545,7 +561,7 @@ describe('KafkaTrigger Node', () => {
 
 	it('should fail activation when the schema registry credential is missing the password', async () => {
 		await expect(
-			testTriggerNode(KafkaTrigger, {
+			testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					credentials: {
@@ -579,7 +595,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should parse JSON message when jsonParseMessage is true', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -611,7 +627,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should include headers when returnHeaders is true', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1,
@@ -661,26 +677,29 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should handle manual trigger mode', async () => {
-		const { emit, manualTriggerFunction } = await testTriggerNode(KafkaTrigger, {
-			mode: 'manual',
-			node: {
-				typeVersion: 1,
-				parameters: {
-					topic: 'test-topic',
-					groupId: 'test-group',
-					useSchemaRegistry: false,
-					options: {
-						parallelProcessing: true,
+		const { emit, manualTriggerFunction } = await testTriggerNode(
+			new KafkaTriggerV1(baseDescription),
+			{
+				mode: 'manual',
+				node: {
+					typeVersion: 1,
+					parameters: {
+						topic: 'test-topic',
+						groupId: 'test-group',
+						useSchemaRegistry: false,
+						options: {
+							parallelProcessing: true,
+						},
 					},
 				},
+				credential: {
+					brokers: 'localhost:9092',
+					clientId: 'n8n-kafka',
+					ssl: false,
+					authentication: false,
+				},
 			},
-			credential: {
-				brokers: 'localhost:9092',
-				clientId: 'n8n-kafka',
-				ssl: false,
-				authentication: false,
-			},
-		});
+		);
 
 		await manualTriggerFunction?.();
 
@@ -696,24 +715,27 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use immediate emit in manual mode even when resolveOffset is onCompletion (v1.3)', async () => {
-		const { emit, manualTriggerFunction } = await testTriggerNode(KafkaTrigger, {
-			mode: 'manual',
-			node: {
-				typeVersion: 1.3,
-				parameters: {
-					topic: 'test-topic',
-					groupId: 'test-group',
-					useSchemaRegistry: false,
-					resolveOffset: 'onCompletion',
+		const { emit, manualTriggerFunction } = await testTriggerNode(
+			new KafkaTriggerV1(baseDescription),
+			{
+				mode: 'manual',
+				node: {
+					typeVersion: 1.3,
+					parameters: {
+						topic: 'test-topic',
+						groupId: 'test-group',
+						useSchemaRegistry: false,
+						resolveOffset: 'onCompletion',
+					},
+				},
+				credential: {
+					brokers: 'localhost:9092',
+					clientId: 'n8n-kafka',
+					ssl: false,
+					authentication: false,
 				},
 			},
-			credential: {
-				brokers: 'localhost:9092',
-				clientId: 'n8n-kafka',
-				ssl: false,
-				authentication: false,
-			},
-		});
+		);
 
 		await manualTriggerFunction?.();
 
@@ -726,26 +748,29 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use immediate emit in manual mode even when parallelProcessing is false (v1.1)', async () => {
-		const { emit, manualTriggerFunction } = await testTriggerNode(KafkaTrigger, {
-			mode: 'manual',
-			node: {
-				typeVersion: 1.1,
-				parameters: {
-					topic: 'test-topic',
-					groupId: 'test-group',
-					useSchemaRegistry: false,
-					options: {
-						parallelProcessing: false,
+		const { emit, manualTriggerFunction } = await testTriggerNode(
+			new KafkaTriggerV1(baseDescription),
+			{
+				mode: 'manual',
+				node: {
+					typeVersion: 1.1,
+					parameters: {
+						topic: 'test-topic',
+						groupId: 'test-group',
+						useSchemaRegistry: false,
+						options: {
+							parallelProcessing: false,
+						},
 					},
 				},
+				credential: {
+					brokers: 'localhost:9092',
+					clientId: 'n8n-kafka',
+					ssl: false,
+					authentication: false,
+				},
 			},
-			credential: {
-				brokers: 'localhost:9092',
-				clientId: 'n8n-kafka',
-				ssl: false,
-				authentication: false,
-			},
-		});
+		);
 
 		await manualTriggerFunction?.();
 
@@ -758,7 +783,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should handle sequential processing when parallelProcessing is false', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.1,
@@ -798,7 +823,7 @@ describe('KafkaTrigger Node', () => {
 	it('should keep binary data when keepBinaryData is enabled in v1.2', async () => {
 		const messageBuffer = Buffer.from('binary-avro-data');
 
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.2,
@@ -839,7 +864,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should not keep binary data in v1.0 and v1.1 even if option is set', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.1,
@@ -879,7 +904,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should convert to string when keepBinaryData is false in v1.2', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.2,
@@ -921,7 +946,7 @@ describe('KafkaTrigger Node', () => {
 		const jsonData = { foo: 'bar', nested: { value: 123 } };
 		const messageBuffer = Buffer.from(JSON.stringify(jsonData));
 
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.2,
@@ -963,7 +988,7 @@ describe('KafkaTrigger Node', () => {
 		const decodedData = { userId: 123, userName: 'test-user' };
 		mockRegistryDecode.mockResolvedValue(decodedData);
 
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.2,
@@ -1007,7 +1032,7 @@ describe('KafkaTrigger Node', () => {
 		const jsonData = { result: 'success', data: [1, 2, 3] };
 		const messageBuffer = Buffer.from(JSON.stringify(jsonData));
 
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.2,
@@ -1047,7 +1072,7 @@ describe('KafkaTrigger Node', () => {
 	it('should keep binary data with returnHeaders enabled', async () => {
 		const messageBuffer = Buffer.from('test-data');
 
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.2,
@@ -1093,7 +1118,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use custom rebalanceTimeout when provided', async () => {
-		await testTriggerNode(KafkaTrigger, {
+		await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1,
@@ -1149,7 +1174,7 @@ describe('KafkaTrigger Node', () => {
 			}),
 		);
 
-		await testTriggerNode(KafkaTrigger, {
+		await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1227,7 +1252,7 @@ describe('KafkaTrigger Node', () => {
 			}),
 		);
 
-		const { close } = await testTriggerNode(KafkaTrigger, {
+		const { close } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1266,7 +1291,7 @@ describe('KafkaTrigger Node', () => {
 		const teardownError = new Error('The group is rebalancing, so a rejoin is needed');
 		mockConsumerStop.mockRejectedValueOnce(teardownError);
 
-		const { close } = await testTriggerNode(KafkaTrigger, {
+		const { close } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1299,7 +1324,7 @@ describe('KafkaTrigger Node', () => {
 		const teardownError = new Error('The coordinator is not aware of this member');
 		mockConsumerDisconnect.mockRejectedValueOnce(teardownError);
 
-		const { close } = await testTriggerNode(KafkaTrigger, {
+		const { close } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1327,7 +1352,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should configure a restartOnFailure callback that vetoes consumer restarts only after close', async () => {
-		const { close } = await testTriggerNode(KafkaTrigger, {
+		const { close } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1362,7 +1387,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should crash the consumer non-retriably when a batch arrives after close', async () => {
-		const { close, emit } = await testTriggerNode(KafkaTrigger, {
+		const { close, emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1395,7 +1420,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should unblock a pending onCompletion execution wait when the trigger is closed', async () => {
-		const { close, emit } = await testTriggerNode(KafkaTrigger, {
+		const { close, emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.3,
@@ -1434,7 +1459,7 @@ describe('KafkaTrigger Node', () => {
 					}),
 			);
 
-			const { close } = await testTriggerNode(KafkaTrigger, {
+			const { close } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					parameters: {
@@ -1475,7 +1500,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use default values for consumer config when options are not provided', async () => {
-		await testTriggerNode(KafkaTrigger, {
+		await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1,
@@ -1504,7 +1529,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should handle batch processing when batchSize > 1', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1542,7 +1567,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should process messages in chunks when batch has more messages than batchSize', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1592,7 +1617,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use fetchMaxBytes and fetchMinBytes when provided', async () => {
-		await testTriggerNode(KafkaTrigger, {
+		await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1,
@@ -1627,7 +1652,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should use partitionsConsumedConcurrently when provided', async () => {
-		await testTriggerNode(KafkaTrigger, {
+		await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1655,7 +1680,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should handle batch processing with sequential processing', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.1,
@@ -1696,7 +1721,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should handle batch processing with JSON parsing', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				parameters: {
@@ -1733,7 +1758,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should handle batch processing with headers', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1,
@@ -1789,7 +1814,7 @@ describe('KafkaTrigger Node', () => {
 	});
 
 	it('should keep binary data in batch processing when keepBinaryData is enabled in v1.2', async () => {
-		const { emit } = await testTriggerNode(KafkaTrigger, {
+		const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 			mode: 'trigger',
 			node: {
 				typeVersion: 1.2,
@@ -1835,7 +1860,7 @@ describe('KafkaTrigger Node', () => {
 
 	describe('version 1.3', () => {
 		it('should use default sessionTimeout and heartbeatInterval', async () => {
-			await testTriggerNode(KafkaTrigger, {
+			await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -1865,7 +1890,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use resolveOffset "immediately" and emit without waiting', async () => {
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -1893,7 +1918,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use resolveOffset "onCompletion" and wait for execution', async () => {
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -1924,7 +1949,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use resolveOffset "onSuccess" and wait for successful execution', async () => {
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -1955,7 +1980,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should return success false when resolveOffset is "onSuccess" and execution fails', async () => {
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -1988,7 +2013,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use sessionTimeout and heartbeatInterval options when provided', async () => {
-			await testTriggerNode(KafkaTrigger, {
+			await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2022,7 +2047,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use resolveOffset "onStatus" and resolve when status matches allowed statuses', async () => {
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2055,7 +2080,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use resolveOffset "onStatus" and fail when status does not match allowed statuses', async () => {
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2090,7 +2115,7 @@ describe('KafkaTrigger Node', () => {
 
 		it('should throw error when resolveOffset is "onStatus" but no statuses are selected', async () => {
 			await expect(
-				testTriggerNode(KafkaTrigger, {
+				testTriggerNode(new KafkaTriggerV1(baseDescription), {
 					mode: 'trigger',
 					node: {
 						typeVersion: 1.3,
@@ -2120,7 +2145,7 @@ describe('KafkaTrigger Node', () => {
 				}
 			});
 
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2212,7 +2237,7 @@ describe('KafkaTrigger Node', () => {
 				}
 			});
 
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2297,7 +2322,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use default auto commit settings (autoCommit true, eachBatchAutoResolve false)', async () => {
-			await testTriggerNode(KafkaTrigger, {
+			await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2325,7 +2350,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should enable eachBatchAutoResolve when option is set', async () => {
-			await testTriggerNode(KafkaTrigger, {
+			await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2356,7 +2381,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use default auto commit settings with onStatus resolveOffset', async () => {
-			await testTriggerNode(KafkaTrigger, {
+			await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2385,7 +2410,7 @@ describe('KafkaTrigger Node', () => {
 		});
 
 		it('should use default auto commit settings with immediately resolveOffset', async () => {
-			await testTriggerNode(KafkaTrigger, {
+			await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2420,7 +2445,7 @@ describe('KafkaTrigger Node', () => {
 				}
 			});
 
-			const { emit } = await testTriggerNode(KafkaTrigger, {
+			const { emit } = await testTriggerNode(new KafkaTriggerV1(baseDescription), {
 				mode: 'trigger',
 				node: {
 					typeVersion: 1.3,
@@ -2498,5 +2523,34 @@ describe('KafkaTrigger Node', () => {
 			expect(mockResolveOffset).toHaveBeenNthCalledWith(2, '1');
 			expect(mockResolveOffset).toHaveBeenNthCalledWith(3, '2');
 		});
+	});
+});
+
+describe('KafkaTrigger (versioned entry point)', () => {
+	const kafkaTrigger = new KafkaTrigger();
+	const expectedDescription = new KafkaTriggerV1(baseDescription).description;
+	const versions = [1, 1.1, 1.2, 1.3];
+
+	it('maps exactly versions 1, 1.1, 1.2, and 1.3 to KafkaTriggerV1', () => {
+		expect(
+			Object.keys(kafkaTrigger.nodeVersions)
+				.map(Number)
+				.sort((a, b) => a - b),
+		).toEqual(versions);
+		for (const version of versions) {
+			expect(kafkaTrigger.nodeVersions[version]).toBeInstanceOf(KafkaTriggerV1);
+		}
+	});
+
+	it('resolves each version to a consistent, correctly-merged description', () => {
+		for (const version of versions) {
+			expect(kafkaTrigger.nodeVersions[version].description).toEqual(expectedDescription);
+		}
+	});
+
+	it('defaults new workflows to version 1.3', () => {
+		expect(kafkaTrigger.description.defaultVersion).toBe(1.3);
+		expect(kafkaTrigger.currentVersion).toBe(1.3);
+		expect(kafkaTrigger.getNodeType()).toBe(kafkaTrigger.nodeVersions[1.3]);
 	});
 });
